allow negative values in LOG_PROGRESS() + var-names

+++ src/Logger.h
@@ -21,7 +21,7 @@ void LOG_ERROR(const char* format, ...);
 */
 void LOG_DOWNLOAD(const char* filename);
 
-void LOG_PROGRESS(unsigned int done,unsigned  int total);
+void LOG_PROGRESS(long done, long total);
 
 #ifdef DEBUG
 #define LOG_DEBUG(fmt, ...) \
